requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.11
django-heroku==0.3.1
django-on-heroku==1.1.2
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
phonenumberslite==8.12.41
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
